Adding new OAuth2 token mod script for the API clients
