+++ config/applications/api-filing-web-client.json.tpl
@@ -30,7 +30,7 @@
 		"issueRefreshToken":true,
 		"accessTokenMayActScript":"[Empty]",
 		"evaluateScopePluginType":"JAVA",
-		"accessTokenModificationScript":"d22f9a0c-426a-4466-b95e-d0f125b0d5fa"
+		"accessTokenModificationScript":"d22f9a0c-426a-4466-b95e-d0f125b0d5fb"
     },
 	"coreOAuth2ClientConfig": {
 		"userpassword": "{API_FILING_CLIENT_PASSWORD}",
